+++ command/delete.go
@@ -13,6 +13,13 @@ func NewDeleteCommand() cli.Command {
 	return cli.Command{
 		Name:  "delete",
 		Usage: "delete a key",
+		Flags: []cli.Flag{
+			cli.BoolFlag{
+				Name:   "recursive, r",
+				Usage:  "resursively delete keys",
+				EnvVar: "",
+			},
+		},
 		Action: func(c *cli.Context) {
 			rawhandle(c, deleteCommandFunc)
 		},
@@ -20,6 +27,8 @@ func NewDeleteCommand() cli.Command {
 }
 
 func deleteCommandFunc(c *cli.Context, client *api.Client) {
+	var err error
+
 	if len(c.Args()) == 0 {
 		fmt.Fprintln(os.Stderr, "Error: Key required")
 		os.Exit(1)
@@ -28,7 +37,12 @@ func deleteCommandFunc(c *cli.Context, client *api.Client) {
 	key := c.Args()[0]
 	kv := client.KV()
 
-	_, err := kv.Delete(key, nil)
+	if c.Bool("recursive") == true {
+		_, err = kv.DeleteTree(key, nil)
+	} else {
+		_, err = kv.Delete(key, nil)
+	}
+
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)