+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Consulctl Changelog
 ===================
 
+## UNRELEASED
+
+  * Add `--recursive` option to `delete`
+
 ## v0.1.0 (April 16, 2015)
 
   * Initial release